\name{addPtsOptNet}
\alias{addPtsOptNet}

\title{
titulo optimNet
}
\description{
descripcion optimNet
}
\usage{
addPtsOptNet(formula, locations, spDatF, fitmodel, n, popSize, generations, xmin, ymin, xmax, ymax, ...)
}

\arguments{
  \item{formula}{
desc var.reg
}
  \item{locations}{
desc coord
}
  \item{spDatF}{
desc newd
}
  \item{fitmodel}{
desc n.vec
}
  \item{n}{
desc n.vec
}
  \item{popSize}{
desc n.vec
}
  \item{generations}{
desc n.vec
}
  \item{xmin}{
desc n.vec
}
  \item{ymin}{
desc n.vec
}
  \item{xmax}{
desc n.vec
}
  \item{ymax}{
desc n.vec
}
  \item{...}{
desc n.vec
}
}
\details{
Detalles
}
\value{
valores retornados
}
\references{
referencias
}
\author{
Alí Santacruz <amsantac@unal.edu.co>
}
\note{
notas
}


\seealso{
ver tambien
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## Ejemplos
library(genalg)

}

\keyword{ spatial }
