\name{optimNet}
\alias{optimNet}

\title{
titulo optimNet
}
\description{
descripcion optimNet
}
\usage{
optimNet(sigma, var.reg, coordenadas, newdata, n.vec)
}

\arguments{
  \item{sigma}{
desc sigma
}
  \item{var.reg}{
desc var.reg
}
  \item{coordenadas}{
desc coord
}
  \item{newdata}{
desc newd
}
  \item{n.vec}{
desc n.vec
}
}
\details{
Detalles
}
\value{
valores retornados
}
\references{
referencias
}
\author{
Alí Santacruz <amsantac@unal.edu.co>
}
\note{
notas
}


\seealso{
ver tambien
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## Ejemplos
library(genalg)

}

\keyword{ spatial }
