\name{melo_0.0-1-package}
\alias{melo_0.0-1-package}
\alias{melo_0.0-1}
\docType{package}
\title{
Rd esto hace
}
\description{
Rd descripcion
}
\details{
\tabular{ll}{
Package: \tab melo_0.0-1\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2011-07-19\cr
License: \tab Rd licencia\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Rd melo

Maintainer: Rd cmelo <yourfault@somewhere.net>
}
\references{
Rd referencias
~~ Literature 
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\keyword{ spatial }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
8*4
}
