\name{geosp-package}
\alias{geosp}
\alias{geosp}
\docType{package}
\title{
Rd esto hace
}
\description{
Rd descripcion
}
\details{
\tabular{ll}{
Package: \tab geosp\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2011-08-02\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Carlos E. Melo <cmelo@udistrital.edu.co>, Alí Santacruz <amsantac@unal.edu.co>

Maintainer: maint <yourfault@somewhere.net>
}
\references{
Rd referencias
~~ Literature 
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\keyword{ spatial }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link{fbr}} ~~
}
\examples{
8*4
}
